Fixing the lazy translations. When translated, nothing was output.
"Normally, you won't have to worry about lazy translations. Just be aware that if you examine an object and it claims to be a django.utils.functional.__proxy__ object, it is a lazy translation. Calling unicode() with the lazy translation as the argument will generate a Unicode string in the current locale."
https://docs.djangoproject.com/en/dev/ref/unicode/#translated-strings